+++ NEWS.md
@@ -5,6 +5,7 @@
 # meteoland 2.0.2
 - Modification of default variables for interpolator_calibration
 - Avoiding interpolation in cells with missing elevation
+- Fixed #21. Now `meteo_complete` works with missing Precipitation
 
 # meteoland 2.0.1
 - Hard deprecation of "sp", "raster" and "rgdal" depedencies, including functions and classes using